Wno Format

+++ main.h
@@ -5,8 +5,8 @@ int _printf(const char *format, ...);
 int _putchar(char c);
 int output_string(char *string);
 int output_decimal(int number);
-long long number_to_base_10(int number, int base);
-long long int power(int base, int a);
+long number_to_base_10(int number, int base);
+long int power(int base, int a);
 int output_binary(int number);
 int output_unsigned_int(unsigned int number);
